Add new OutputPreview sample to public APDFL samples repository
OutputPreview.cpp is related to the CreateSeparations.cpp sample.

The sample demonstrates how to create color separations for spot color images, but allows for the combination of these plates into a single DeviceN color image, which displays the true color representation of the page. It generates a series of image files allowing a user to preview the colorants involved.