Add new DotNET Core C# samples to repository
Datalogics is providing the Microsoft DotNETCore framework with Adobe
PDF Library 18. This new platform provides for cross-platform
development in C# for Windows, Linux, or Mac.

The C# samples provided with DotNET Core mostly match those provided in
the existing DotNET interface, but the Forms samples are not included,
and the samples DotNETViewer and PDFObjectExplorer are not included as
well.