Update to github repository to improve usefulness
README file revised to better explain the purpose of this repository.

With this pull request, solution, project, and user files related to C++ and C# sample programs
are removed.

The sample program files provided with this repository are for
reference only. You can't build and run these programs from this site.
To run these sample programs, you need to install Adobe PDF Library.

These supporting files were added to this repository when it was created. They are removed
here because they are not needed.

We seek to avoid confusing users of this repository.