Update github repository, .NET samples
WIth this pull request, solution and project files related to the C#
sample programs for the .NET Interface are removed.

The sample C# program files provided with this repository are for
reference only. You can't run these programs from this site.  You need
to install Adobe PDF Library to run these samples. We are removing them
because they are not needed, and to avoid confusing users of this
repository.

This commit is for Annotations and Content Creation.